added Captive Portal functionality
In this commit I changed the httpsetup and httploop functions and added webHandleCaptivePortalWifiConfig to send the wifi config page when a device connects to the softAP. 
This will simplify the innitial wifi setup. The usercan just scan the QR code and a webpage will automatically appear without the need to enter the IP address.